RaiiList eases handling dynamic amount of frames
The RaiiList struct represents a dynamic list which is erased when
getting out of the scope. So it is possible to have a dynamic list
of raii elements that will be properly destructed when going out of
scope.

This is used in the recvFrames function which will returns all frames
in a message without knowing priorly the amount of frames.